magit-mode-quit-window: new function
Add new function `magit-mode-quit-window' as an alternative to
`magit-restore-window-configuration'.  Use the new function as
the new default value of `magit-bury-buffer-function'.

Restoring the previous window configuration often is not desirable as
it would be better to keep windows and buffers created or visited
later. Also the point is stored in window configurations and restoring
that is often very undesirable.